+++ examples/published/tutorial_werthmuller-2017-TLE.py
@@ -3,7 +3,7 @@ empymod-tutorial: Werthmüller, 2017, The Leading Edge
 =====================================================
 
 See the notebooks in the repo `empymod/article-tle2017
-<https://github.com/empymod/article-tle2017>`_ for:
+<https://github.com/emsig/article-tle2017>`_ for:
 
     Werthmüller, D., 2017, **Getting started with controlled-source
     electromagnetic 1D modeling**: The Leading Edge, 36, 352-355; doi: